[fix] routes after refactoring /experimental components
