+++ hypergan/gans/experimental/multi_generator_gan.py
@@ -21,10 +21,9 @@ import tensorflow as tf
 import hypergan as hg
 
 from hypergan.gan_component import ValidationException, GANComponent
-from .base_gan import BaseGAN
+from ..base_gan import BaseGAN
 
 from hypergan.distributions.uniform_distribution import UniformDistribution
-from hypergan.trainers.multi_step_trainer import MultiStepTrainer
 
 class MultiGeneratorGAN(BaseGAN):
     """
