+++ hypergan/samplers/static_batch_sampler.py
@@ -1,7 +1,7 @@
 from hypergan.samplers.base_sampler import BaseSampler
 import numpy as np
 import tensorflow as tf
-from hypergan.train_hooks.imle_train_hook import IMLETrainHook
+from hypergan.train_hooks.experimental.imle_train_hook import IMLETrainHook
 
 class StaticBatchSampler(BaseSampler):
     def __init__(self, gan, samples_per_row=8):
